Run keda in a Docker container
Update installation instructions to create and use a Docker image.

Use a jessie image and install django from wheezy. Since both releases
are not supported any more we have to enable the archives repos [1].

Use /keda as working dir. Use dumb-init [2] as PID 1 and have it run a
start.sh script that initializes/migrates the db, and then starts
gunicorn (using the debian wrapper) and nginx and sleeps forever. Link
logfiles to stdout/stderr [2] so that we can see nginx and gunicorn logs
as container logs.

To start the container you need to:

* bind mount the DB file under /keda/db.sqlite3
* bind mount the SSL certificate under /etc/ssl/private/keda.key and
/etc/ssl/certs/keda.pem

Fix nginx config to set x-forwarded-proto otherwise we were getting
redirected to http and sending HTTP requests to HTTPS port.

[1] https://wiki.debian.org/DebianWheezy
[2] https://github.com/Yelp/dumb-init
[3] https://stackoverflow.com/questions/22541333/have-nginx-access-log-and-error-log-log-to-stdout-and-stderr-of-master-process

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>